Migrate :appcompat-theme tests to Robolectric
