+++ appcompat-theme/src/sharedTest/res/font/rubik_300.ttf
