+++ appcompat-theme/src/sharedTest/res/font/rubik_500.ttf
