+++ appcompat-theme/src/test/AndroidManifest.xml
@@ -1,0 +1,34 @@
+<!--
+  ~ Copyright 2020 The Android Open Source Project
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      https://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.google.accompanist.appcompattheme">
+
+    <application>
+        <activity
+            android:name=".LightAppCompatActivity"
+            android:theme="@style/Theme.AppCompatThemeTest" />
+
+        <activity
+            android:name=".DarkAppCompatActivity"
+            android:theme="@style/Theme.AppCompatThemeTest" />
+
+        <activity
+            android:name=".NotAppCompatActivity"
+            android:theme="@android:style/Theme.Material" />
+    </application>
+
+</manifest>
